LayoutTests/inspector/debugger/raw-source-code-expected.txt

Tests RawSourceCode class.


Running: testScriptWithoutResource
UISourceCode: foo.js
UISourceCode is content script: true
Mime type: text/javascript
Script source: <script source>

Running: testHTMLWithPendingResource
Adding first script for pending request.
UISourceCode: index.html
UISourceCode is content script: false
Mime type: text/html
Script source:   <script><script source 1></script>
Adding second script for pending request.
UISourceCode: index.html
UISourceCode is content script: false
Mime type: text/html
Script source:   <script><script source 1></script> <script><script source 2></script>
Finishing request.
UISourceCode: index.html
UISourceCode is content script: false
Mime type: text/html
Script source: <resource content>

Running: testHTMLWithFinishedResource
UISourceCode: index.html
UISourceCode is content script: false
Mime type: text/html
Script source: <resource content>
